requirements.txt
python-dotenv==1.1.1
loguru==0.7.3
fastapi==0.118.0
uvicorn==0.37.0
pydantic-settings==2.5.2
redis==7.0.0
httpx==0.28.1
geoalchemy2==0.18.0
shapely==2.1.2
asyncpg==0.29.0